refactor code, use folder names to determine parser
